bugfix: Fix the issues of toast messages in client and project management (#591)
This PR introduces:
In Clients, during Add and Update toast message is correctly displayed.
In Contracts, also toast message is dispalyed correctly.
In Clients, when logo is uploaded remove icon position is fixed.
In Contracts and Clients, for attachments file format is added an for
logo also file format is added.
In Employees, in Documents tab file format position is corrected above
the submit button.